+++ renderer/page.go
@@ -46,7 +46,9 @@ func CanDisplay(res *gemini.Response) bool {
 	return err == nil && enc != nil
 }
 
-func MakePage(url string, res *gemini.Response, width int) (*structs.Page, error) {
+// MakePage creates a formatted, rendered Page from the given network response and params.
+// You must set the Page.Width value yourself.
+func MakePage(url string, res *gemini.Response, width, leftMargin int) (*structs.Page, error) {
 	if !CanDisplay(res) {
 		return nil, errors.New("not valid content for a Page")
 	}
@@ -76,17 +78,27 @@ func MakePage(url string, res *gemini.Response, width int) (*structs.Page, error
 	}
 
 	if mediatype == "text/gemini" {
-		rendered, links := RenderGemini(utfText, width)
+		rendered, links := RenderGemini(utfText, width, leftMargin)
 		return &structs.Page{
 			Url:     url,
+			Raw:     utfText,
 			Content: rendered,
 			Links:   links,
 		}, nil
 	} else if strings.HasPrefix(mediatype, "text/") {
 		// Treated as plaintext
+
+		// Add left margin
+		var shifted string
+		lines := strings.Split(utfText, "\n")
+		for i := range lines {
+			shifted += strings.Repeat(" ", leftMargin) + lines[i] + "\n"
+		}
+
 		return &structs.Page{
 			Url:     url,
-			Content: utfText,
+			Raw:     utfText,
+			Content: shifted,
 			Links:   []string{},
 		}, nil
 	}