+++ structs/structs_test.go
@@ -9,8 +9,9 @@ import (
 func TestSize(t *testing.T) {
 	p := Page{
 		Url:     "12345",
+		Raw:     "12345",
 		Content: "12345",
 		Links:   []string{"1", "2", "3", "4", "5"},
 	}
-	assert.Equal(t, 15, p.Size(), "sizes should be equal")
+	assert.Equal(t, 20, p.Size(), "sizes should be equal")
 }