+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Pages are rewrapped dynamically, whenever the terminal size changes (#33)
+
 ## [1.2.0] - 2020-07-02
 ### Added
 - Alt-Left and Alt-Right for history navigation (#23)